requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
et-xmlfile==1.1.0
idna==3.2
numpy==1.21.1
openpyxl==3.0.7
pandas==1.3.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
